move PeriodIndex comparisons, implement PeriodArray constructor (#21798)
